Add CLC ablation gates, AWQ timing, GPTQ dtype fix
- SmartFlipConfig.enable_knee_mask / enable_max_flip_cap; main.py CLI flags --disable-knee-mask / --disable-max-flip-cap wired through pipeline.
- AWQ smart_flip apply timing recorded in AWQQuantizerXL.timing_stats, surfaced via metadata.
- GPTQ dead_mask copy casts to corrected dtype to avoid Half/Float Index put crash.
- scripts/bash/ablation/run_llama31_awq_ablation.sh runs best / no_k / no_f / all variants with k=0.02, f=0.01.

+++ src/quantization/gptq.py
@@ -267,7 +267,7 @@ class GPTQ:
             quant_state = self._build_quant_state(original_weights, pre_round, integer_weights, scale, zero)
             corrected_weights, outlier_percent, flip_stats = self.smart_flip.apply(quant_state, act_mean)
             if dead_mask is not None and torch.any(dead_mask):
-                corrected_weights[:, dead_mask] = quant_weights[:, dead_mask]
+                corrected_weights[:, dead_mask] = quant_weights[:, dead_mask].to(corrected_weights.dtype)
 
         if self.bias_correction is not None and isinstance(self.layer, nn.Linear):
             bias_delta = self.bias_correction.compute_bias_delta(
